DjangoPJ/MapasApp/requirements.txt
amqp==2.5.2
astroid==2.3.3
billiard==3.6.2.0
branca==0.3.1
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==2.2
django-filter==2.2.0
djangorestframework==3.11.0
folium==0.10.1
gunicorn==20.0.4
idna==2.8
isort==4.3.21
Jinja2==2.11.1
kombu==4.6.7
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.18.1
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
PyMySQL==0.9.3
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
